Move MainApp to the main directory.
