+++ frontend/index.html
@@ -14,6 +14,6 @@
 </head>
 <body>
 <div id="root"></div>
-<script src="src/main/main.tsx" type="module"></script>
+<script src="src/main/index.tsx" type="module"></script>
 </body>
 </html>
