+++ frontend/src/main/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { createRoot } from 'react-dom/client'
-import App from './app/MainApp'
+import App from './MainApp'
 
 const container = document.getElementById('root')!
 const root = createRoot(container)
